+++ client/index.js
@@ -1,13 +1,19 @@
-export { default as AccountsOverview } from './screens/AccountsOverview'
-export { default as Home } from './screens/Home'
-export { default as Initial } from './screens/Initial'
-export { default as Link } from './screens/Link'
-export { default as Login } from './screens/Login'
-export { default as Quiz } from './screens/Quiz'
-export { default as Result } from './screens/Result'
-export { default as Signup } from './screens/Signup'
-export { default as Accounts } from './screens/Accounts'
-export { default as Budget } from './screens/Budget'
-export { default as Profile } from './screens/Profile'
-export { default as Navbar } from './screens/Navbar'
-export { default as Main } from './screens/Main'
+// Screens
+export { default as Home } from "./screens/Home";
+export { default as Initial } from "./screens/Initial";
+export { default as Link } from "./screens/Link";
+export { default as Login } from "./screens/Login";
+export { default as Quiz } from "./screens/Quiz";
+export { default as Result } from "./screens/Result";
+export { default as Signup } from "./screens/Signup";
+export { default as Navbar } from "./screens/Navbar";
+export { default as Main } from "./screens/Main";
+
+// Accounts
+export { default as Accounts } from "./screens/Accounts";
+
+// Budget
+export { default as Budget } from "./screens/Budget";
+
+// Profile
+export { default as Profile } from "./screens/Profile";
