+++ client/screens/Budget/index.js
@@ -1,9 +1,8 @@
-import React from 'react';
+import React from "react";
+import { Text, View } from "react-native";
 
 const Budget = () => {
-  return (
-    <View />
-  )
-}
+  return <View />;
+};
 
 export default Budget;
