+++ client/screens/Navbar.js
@@ -1,24 +1,29 @@
-import React from 'react';
-import { Text, View } from 'react-native';
-import {Home, Accounts, Budget, Profile} from '../'
-import {createMaterialBottomTabNavigator} from "react-navigation-material-bottom-tabs";
+import React from "react";
+import { Text, View } from "react-native";
+// import { Home, Accounts, Budget, Profile } from "../";
+import Home from "./Home";
+import Budget from "./Budget";
+import Accounts from "./Accounts";
 
+import { createMaterialBottomTabNavigator } from "react-navigation-material-bottom-tabs";
 
 const Navbar = () => {
-  return (<TabNav />)
-}
+  return <TabNav />;
+};
 
-const TabNav = createMaterialBottomTabNavigator({
-  Home: { screen: Home },
-  Budget: { screen: Budget },
-  Accounts: { screen: Accounts },
-  Profile: { screen: Profile },
-}, {
-  initialRouteName: 'Home',
-  activeTintColor: '#f0edf6',
-  inactiveTintColor: '#3e2465',
-  barStyle: { backgroundColor: '#694fad' },
-});
+const TabNav = createMaterialBottomTabNavigator(
+  {
+    Home: { screen: Home },
+    Budget: { screen: Budget },
+    Accounts: { screen: Accounts }
+    // Profile: { screen: Profile },
+  },
+  {
+    initialRouteName: "Home",
+    activeTintColor: "#efcdba",
+    inactiveTintColor: "#FFFFFF",
+    barStyle: { backgroundColor: "#7CA5B8" }
+  }
+);
 
 export default Navbar;
-
