+++ client/screens/Profile/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Text, View } from 'react-native';
 
 const Profile = () => {
   return (
