+++ package.json
@@ -28,6 +28,7 @@
     "react-native-elements": "^0.19.1",
     "react-native-plaid-link": "^1.3.5",
     "react-native-slider": "^0.11.0",
+    "react-native-vector-icons": "^4.6.0",
     "react-navigation": "^2.3.1",
     "react-navigation-material-bottom-tabs": "^0.3.0",
     "react-redux": "^5.0.7",
